requirements.txt
numpy
yfinance
pandas
langgraph
langchain-google-genai
python-dotenv
streamlit
plotly
